move zones methods to their own file
